Inconsistent use of user-supplied data and data from the database in the subscription form in some cases.
The firstrun flag was used erroneously. This caused the form to
always include data from the user record which were then being
overwritten by user data. Under some conditions the overwrite did
not work. This could be a minor nuissance (business information
not being update, for instance) or an outright validation failure
which prevented the subscription from going on. This commit
addresses the inconsistency in the use of the firstrun flag in
the session.

Moreover, we address a problem with the validation data reset when
the firstrun flag is set. While we were resetting the StateData
object used internally by the Subscribe model we did not reset the
user state saved by Joomla itself. As a result, the very next thing
that happened was that the erroneous information from the global
Joomla user state was repopulating the StateData object with the
exact data we wanted to remove in the first place. Awkward.

Finally, we added a URL parameter reset=1 in the Levels view
(subscription page) which allows us to forcibly reset the saved
state data, i.e. set the firstrun flag in the session. When the
user submits a form with validation errors we display a message
with a link containing the reset=1 URL parameter, allowing them to
start over the subscription process. This should not be necessary
but it's a good fallback in case our clients end up having stale
session data for any reason we cannot currently imagine. Better
be safe than sorry!