Improved use of timezones throughout the component
Close gh-295

We added user timezone awareness in the following areas:

* Message parsing which is used in emails, thank you pages etc
* The subscription expiration module
* Log messages when the subscription expiration plugin runs

Please note that in the message parsing (Helper\Message) we pass the
user ID found in the subscription. This allows us to tailor the
message for the time zone selected by the user as opposed to the
time zone of the currently logged in user (which is typically GMT
since these things run from the CLI).