Update the tests for the removal of the dev-only vendor folder
We used to have a vendor folder in the repo's root with just development
requirements. We have done away with it. The composer.json only contains
production requirements.

You should now use PHPUnit installed separately, e.g. globally on your
user account using `composer global require phpunit/phpunit "^7.5"`.

As a result the check for the `vendor` folder is no longer valid and had
to be removed.