Reworking the My Subscriptions page
Don't show Renew for fixed expiration date subscriptions UNLESS the
level's expiration date has been updated to be in the future of the
existing subscription's expiration date AND the level's fixed expiration
date is in the absolute future from now.

This makes sense with public beta subscriptions. Let's say the original
planning was to have the beta expire on May 1st but around April we see
that actually need more time. We can update the level's fixed expiration
to, say, end of July. We need to allow existing beta susbcribers to
extend their beta access, therefore they need to see the Renew button.
However, after the end of July the beta is closed. We don't want them to
see the Renew button anymore.